+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.0 - 2024-05-04
+
+- Support for custom Lambda code
+- Now requires `ecr_image_filter_lambda_asset_path` and `soci_index_generator_lambda_asset_path` to be provided
+- A new module `copy-lambda-code` is provided to copy the Lambda code from the AWS-IA bucket to your own S3 bucket
+- Rename `deployment_assets_bucket_name` -> `deployment_assets_bucket`
+
 ## 0.0.1 - 2024-05-04
 
 - Initial commit